Adds the ability to configure AileronAssist to control a standard plane, a v-tail, a delta wing, or a helicopter.
(When in helicopter mode, it also disables navigation.  I did this by limiting the state machine.  Comments are welcome on better ways to do this!)

But please note that this isn't tested yet!  So please don't fly a plane based on this code without doing tons of testing of your own.  :)